--- homeworks/hw1/read_me
@@ -1,1 +1,0 @@
-hello
